fix: avoid unnecessary casting

+++ ga/ga.go
@@ -46,12 +46,12 @@ type PriorityWeights struct {
 }
 
 type Weights struct {
-	RowBonus     int64
-	PosBonus     int64
-	MedPenalty   int64
-	FriendBonus  int64
-	EnemyPenalty int64
-	PrefBonus    int64
+	RowBonus     int
+	PosBonus     int
+	MedPenalty   int
+	FriendBonus  int
+	EnemyPenalty int
+	PrefBonus    int
 }
 
 func calculateWeights(config ClassConfig, pw PriorityWeights) Weights {
@@ -62,14 +62,13 @@ func calculateWeights(config ClassConfig, pw PriorityWeights) Weights {
 		BASE_HARD   = 10000
 	)
 
-	wFill := int64(pw.Fill * float64(BASE_FILL))
-	wPref := int64(pw.Preferences * float64(BASE_PREF))
-	wFriend := int64(pw.Friends * float64(BASE_FRIEND))
-	wMed := int64(pw.Medical * float64(BASE_HARD))
-	wEnemy := int64(pw.Enemies * float64(BASE_HARD))
-
+	wFill := int(pw.Fill * float64(BASE_FILL))
+	wPref := int(pw.Preferences * float64(BASE_PREF))
+	wFriend := int(pw.Friends * float64(BASE_FRIEND))
+	wMed := int(pw.Medical * float64(BASE_HARD))
+	wEnemy := int(pw.Enemies * float64(BASE_HARD))
 	return Weights{
-		RowBonus: int64(config.Columns) * wFill * 10,
+		RowBonus: int(config.Columns) * wFill * 10,
 		PosBonus: wFill,
 
 		PrefBonus:    wPref,
@@ -116,7 +115,7 @@ func buildSocialMap(req Request) (SocialMap, SocialMap) {
 	return friends, enemies
 }
 
-func fitness(seating []int, students []Student, preferences, forbidden [][]int, config ClassConfig, w Weights) (int64, []int) {
+func fitness(seating []int, students []Student, preferences, forbidden [][]int, config ClassConfig, w Weights) (int, []int) {
 	ignored := make([]int, 0)
 	score := 0
 	for i, studentIndex := range seating {
@@ -126,15 +125,15 @@ func fitness(seating []int, students []Student, preferences, forbidden [][]int, 
 		student := students[studentIndex]
 		row := i / config.Columns
 		col := i % config.Columns
-		score += (config.Rows - row) * int(w.RowBonus)
-		score += (config.Columns - col) * int(w.PosBonus)
+		score += (config.Rows - row) * w.RowBonus
+		score += (config.Columns - col) * w.PosBonus
 		if (len(student.PreferredRows) > 0 && contains(student.PreferredRows, row)) || len(student.PreferredColumns) > 0 && contains(student.PreferredColumns, col) {
-			score += int(w.PrefBonus) * config.Rows
+			score += w.PrefBonus * config.Rows
 			ignored = append(ignored, student.ID)
 		}
 
 		if len(student.MedicalPreferredColumns) > 0 && !contains(student.MedicalPreferredColumns, col) || len(student.MedicalPreferredRows) > 0 && !contains(student.MedicalPreferredRows, row) {
-			score -= int(w.MedPenalty)
+			score -= w.MedPenalty
 		}
 
 	}
@@ -153,20 +152,20 @@ func fitness(seating []int, students []Student, preferences, forbidden [][]int, 
 			i2ID := students[seating[i+1]].ID
 			for _, pref := range preferences {
 				if (pref[0] == i1ID && pref[1] == i2ID) || (pref[0] == i2ID && pref[1] == i1ID) {
-					score += int(w.FriendBonus)
+					score += w.FriendBonus
 				} else if pref[0] == i1ID || pref[1] == i1ID || pref[0] == i2ID || pref[1] == i2ID {
 					ignored = append(ignored, i1ID, i2ID)
 				}
 			}
 			for _, forb := range forbidden {
 				if (forb[0] == i1ID && forb[1] == i2ID) || (forb[0] == i2ID && forb[1] == i1ID) {
-					score -= int(w.EnemyPenalty)
+					score -= w.EnemyPenalty
 					ignored = append(ignored, i1ID, i2ID)
 				}
 			}
 		}
 	}
-	return int64(score), ignored
+	return score, ignored
 }
 
 func CrossOver(parent1, parent2 []int) []int {
@@ -215,7 +214,7 @@ func SwapMutation(seating []int) []int {
 	return seat
 }
 
-func tournamentSelection(population [][]int, scores []int64, k int) []int {
+func tournamentSelection(population [][]int, scores []int, k int) []int {
 	bestIdx := -1
 	for i := 0; i < k; i++ {
 		randIdx := rand.Intn(len(population))
@@ -225,7 +224,7 @@ func tournamentSelection(population [][]int, scores []int64, k int) []int {
 	}
 	return population[bestIdx]
 }
-func RunGA(req Request) ([]Response, int64, []int) {
+func RunGA(req Request) ([]Response, int, []int) {
 	N := req.ClassConfig.Columns * req.ClassConfig.Rows
 	popSize, generations := req.PopSize, req.Generations
 	weights := calculateWeights(req.ClassConfig, req.PriorityWeights)
@@ -234,7 +233,7 @@ func RunGA(req Request) ([]Response, int64, []int) {
 		population[i] = rand.Perm(N)
 	}
 	for gen := 0; gen < generations; gen++ {
-		scores := make([]int64, popSize)
+		scores := make([]int, popSize)
 		ignored := make([][]int, popSize)
 		for i, seat := range population {
 			scores[i], ignored[i] = fitness(seat, req.Students, req.Preferences, req.Forbidden, req.ClassConfig, weights)
